Refactor StaticResponseAction to be a MessageWithReplyMarkupBuilder
